feat: refactor result handling and improve API responses; update subject references, enhance error handling, and streamline dashboard components for better user experience
