+++ src/app/prof/layout.tsx
@@ -15,7 +15,7 @@ import {
 import LogoutButton from "@/components/Logout";
 
 const navigation = [
-  { name: "Dashboard", href: "/prof/dashboard", icon: Home },
+  { name: "Dashboard", href: "/prof", icon: Home },
   { name: "Marks Entry", href: "/prof/marks/selection", icon: BookOpen },
   { name: "Attendance", href: "/prof/attendance/selection", icon: BookOpen },
   { name: "Results", href: "/prof/results/selection", icon: BarChart2 },
