+++ src/app/prof/results/view-result/page.tsx
@@ -13,6 +13,7 @@ interface Result {
   exam: string;
   marksObtained: number;
   subject: string;
+  student: string | { _id: string }; // for populated or plain
 }
 
 const EXAMS = [
@@ -36,14 +37,24 @@ export default function ViewResultPage() {
 
   useEffect(() => {
     const fetchBatchData = async () => {
-      if (!batchCode || !subjectName || !semester) return;
+      if (!subjectName || !batchCode || !semester) {
+        setError("Missing required query parameters.");
+        return;
+      }
+
+      const modifiedBatchCode = batchCode.replace("-", `${semester}0`);
+
+      console.log("Fetching with:", { subjectName, modifiedBatchCode, semester });
       setLoading(true);
+
       try {
+        // Fetch student list
         const res = await fetch("/api/teaching-assignments/fetch-batch", {
           method: "POST",
           headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ batchCode, subjectName, semester }),
+          body: JSON.stringify({batchCode, subjectName, semester }),
         });
+
         const data = await res.json();
         if (data.success) {
           setStudents(data.students);
@@ -52,43 +63,66 @@ export default function ViewResultPage() {
           setError(data.error || "Failed to fetch students");
         }
 
-        // Fetch results for all students in this subject using the new endpoint
+        // Fetch marks
         const resultRes = await fetch("/api/result/fetch-marks", {
           method: "POST",
           headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ subjectName, batchCode, sem: semester }),
+          body: JSON.stringify({ subjectName, batchCode: modifiedBatchCode, sem: semester }),
         });
-        const resultData = await resultRes.json();
-        setDebugResults(resultData); // Debug: store raw results
+
+        const rawText = await resultRes.text();
+        if (!rawText) {
+          throw new Error("Empty response from fetch-marks API.");
+        }
+
+        let resultData: any;
+        try {
+          resultData = JSON.parse(rawText);
+        } catch (jsonErr) {
+          console.error("JSON parse error:", rawText);
+          throw new Error("Invalid JSON response from server.");
+        }
+
+        console.log("Fetched results:", resultData);
+        setDebugResults(resultData);
+
         const grouped: Record<string, Result[]> = {};
         if (resultData?.results) {
           for (const r of resultData.results) {
-            if (!grouped[r.student]) grouped[r.student] = [];
-            grouped[r.student].push(r);
+            const studentId = typeof r.student === "string" ? r.student : r.student._id;
+            if (!grouped[studentId]) grouped[studentId] = [];
+            grouped[studentId].push(r);
           }
         }
+
         setResults(grouped);
       } catch (err: any) {
+        console.error("Error fetching:", err);
         setError(err.message || "Unknown error");
       } finally {
         setLoading(false);
       }
     };
+
     fetchBatchData();
   }, [subjectName, batchCode, semester]);
 
-  if (loading) return <div className="text-white text-center mt-10">Loading...</div>;
-  if (error) return <div className="text-red-500 text-center mt-10">Error: {error}</div>;
+  if (loading)
+    return <div className="text-white text-center mt-10">Loading...</div>;
+  if (error)
+    return <div className="text-red-500 text-center mt-10">Error: {error}</div>;
 
   return (
     <div className="bg-black text-white min-h-screen p-6 space-y-8">
       <h2 className="text-3xl font-bold text-white/90 mb-6">📊 View Results</h2>
+
       {/* Debug output */}
       {debugResults && (
         <pre className="bg-gray-900 text-green-400 p-4 mb-4 rounded text-xs overflow-x-auto max-h-64">
           {JSON.stringify(debugResults, null, 2)}
         </pre>
       )}
+
       <div className="overflow-x-auto rounded-lg border border-gray-800">
         <table className="min-w-full bg-gray-950 text-gray-300 text-sm">
           <thead className="bg-gray-800 text-gray-300 uppercase">
@@ -96,7 +130,12 @@ export default function ViewResultPage() {
               <th className="px-6 py-3 border-b border-gray-700 text-left">Roll No</th>
               <th className="px-6 py-3 border-b border-gray-700 text-left">Name</th>
               {EXAMS.map((exam) => (
-                <th key={exam.key} className="px-6 py-3 border-b border-gray-700 text-center">{exam.label}</th>
+                <th
+                  key={exam.key}
+                  className="px-6 py-3 border-b border-gray-700 text-center"
+                >
+                  {exam.label}
+                </th>
               ))}
             </tr>
           </thead>
@@ -104,14 +143,21 @@ export default function ViewResultPage() {
             {students.map((student) => {
               const studentResults = results[student._id] || [];
               return (
-                <tr key={student._id} className="border-b border-gray-800 transition hover:bg-gray-800">
+                <tr
+                  key={student._id}
+                  className="border-b border-gray-800 transition hover:bg-gray-800"
+                >
                   <td className="px-6 py-3 whitespace-nowrap text-left">{student.roll}</td>
                   <td className="px-6 py-3 whitespace-nowrap text-left">{student.name}</td>
                   {EXAMS.map((exam) => {
                     const res = studentResults.find((r) => r.exam === exam.key);
                     return (
                       <td key={exam.key} className="px-6 py-3 text-center">
-                        {res ? res.marksObtained : <span className="text-yellow-400/80 font-semibold">Pending</span>}
+                        {res ? (
+                          res.marksObtained
+                        ) : (
+                          <span className="text-yellow-400/80 font-semibold">Pending</span>
+                        )}
                       </td>
                     );
                   })}
