+++ src/models/exams/result.model.ts
@@ -3,7 +3,7 @@ import mongoose, { Schema, Document, models } from "mongoose";
 interface IResult extends Document {
   student: mongoose.Types.ObjectId;
   exam: string;
-  subject: mongoose.Types.ObjectId;
+  subject: string;
   marksObtained: number;
   sem: string;
   batchCode: string; 
@@ -13,7 +13,7 @@ const ResultSchema = new Schema<IResult>(
   {
     student: { type: Schema.Types.ObjectId, ref: "Student", required: true },
     exam: { type: String, required: true },
-    subject: { type: Schema.Types.ObjectId, ref: "Subject", required: true },
+    subject: { type: String, required: true },
     marksObtained: { type: Number, required: true },
     sem: { type: String, required: true}, 
     batchCode: { type: String, required: true },
